rfunctions-package {rfunctions}	R Documentation

A set of R functions that I find useful

Description

R functions that are useful or are fast. Most of the implementation is in C++, utilizing RcppEigen

Details

Package: 	 rfunctions

Type: 	 Package

Version: 	 1.0

Date: 	 2014-03-15

License: 	 GPL (>= 2)

Depends: 	 Rcpp, RcppEigen

oemfitFast is the main OEM function. It takes x and y as arguments for data and produces the regularization path over a grid of values for lambda. It accepts x as either a 'matrix' or 'sparseMatrix'

Author(s)

Jared Huling

Maintainer: Jared Huling <jaredhuling [ at ] gmail [ dot ] com>

Examples

n.obs <- 1e5
n.vars <- 150

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)

## compute X'X
xpx <- crossprodcpp(x)

weights <- runif(n.obs)

## compute X'WX
xpwx <- crossprodcpp(x, weights)

[Package rfunctions version 0.1 Index]